--- docs/source/api_reference/framework/event.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium.framework.event
