+++ docs/source/api_reference/framework/event/index.rst
@@ -1,0 +1,11 @@
+================
+Event Management
+================
+
+.. automodule:: vivarium.framework.event
+
+.. toctree::
+   :maxdepth: 1
+   :glob:
+
+   *
