+++ docs/source/api_reference/framework/event/interface.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium.framework.event.interface
