+++ docs/source/api_reference/framework/event/manager.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium.framework.event.manager
