+++ docs/source/api_reference/framework/logging/index.rst
@@ -1,3 +1,7 @@
+==================
+Logging Management
+==================
+
 .. automodule:: vivarium.framework.logging
 
 .. toctree::
