+++ docs/source/api_reference/framework/lookup/index.rst
@@ -1,3 +1,7 @@
+=======================
+Lookup Table Management
+=======================
+
 .. automodule:: vivarium.framework.lookup
 
 .. toctree::
