+++ src/vivarium/framework/event/__init__.py
@@ -1,0 +1,2 @@
+from vivarium.framework.event.interface import EventInterface
+from vivarium.framework.event.manager import Event, EventChannel, EventManager
