+++ src/vivarium/framework/event/manager.py
@@ -1,13 +1,11 @@
 """
-============================
-The Vivarium Event Framework
-============================
+=============
+Event Manager
+=============
 
 ``vivarium`` constructs and manages the flow of :ref:`time <time_concept>`
 through the emission of regularly scheduled events. The tools in this module
-manage the relationships between event emitters and listeners and provide
-an interface for user :ref:`components <components_concept>` to register
-themselves as emitters or listeners to particular events.
+manage the relationships between event emitters and listeners.
 
 The :class:`EventManager` maintains a mapping between event types and channels.
 Each event type (and event types must be unique so event type is equivalent to
@@ -15,16 +13,11 @@ event name, e.g., ``time_step_prepare``) corresponds to an
 :class:`EventChannel`, which tracks listeners to that event in prioritized
 levels and passes on the event to those listeners when emitted.
 
-The :class:`EventInterface` is exposed off the :ref:`builder <builder_concept>`
-and provides two methods: :func:`get_emitter <EventInterface.get_emitter>`,
-which returns a callable emitter for the given event type and
-:func:`register_listener <EventInterface.register_listener>`, which adds the
-given listener to the event channel for the given event. This is the only part
-of the event framework with which client code should interact.
-
 For more information, see the associated event :ref:`concept note <event_concept>`.
 
 """
+
+
 from __future__ import annotations
 
 from collections.abc import Callable
@@ -35,7 +28,7 @@ from typing import TYPE_CHECKING, Any
 import pandas as pd
 
 from vivarium.framework.lifecycle import ConstraintError, lifecycle_states
-from vivarium.manager import Interface, Manager
+from vivarium.manager import Manager
 from vivarium.types import ClockStepSize, ClockTime
 
 if TYPE_CHECKING:
@@ -281,71 +274,3 @@ class EventManager(Manager):
 
     def __repr__(self) -> str:
         return "EventManager()"
-
-
-class EventInterface(Interface):
-    """The public interface for the event system."""
-
-    def __init__(self, manager: EventManager):
-        self._manager = manager
-
-    def get_emitter(
-        self, event_name: str
-    ) -> Callable[[pd.Index[int], dict[str, Any] | None], Event]:
-        """Gets an emitter for a named event.
-
-        Parameters
-        ----------
-        event_name
-            The name of the event the requested emitter will emit.
-            Users may provide their own named events by requesting an emitter
-            with this function, but should do so with caution as it makes time
-            much more difficult to think about.
-
-        Returns
-        -------
-            An emitter for the named event. The emitter should be called by
-            the requesting component at the appropriate point in the simulation
-            lifecycle.
-        """
-        return self._manager.get_emitter(event_name)
-
-    def register_listener(
-        self, event_name: str, listener: Callable[[Event], None], priority: int = 5
-    ) -> None:
-        """Registers a callable as a listener to a events with the given name.
-
-        The listening callable will be called with a named ``Event`` as its
-        only argument any time the event emitter is invoked from somewhere in
-        the simulation.
-
-        The framework creates the following events and emits them at different
-        points in the simulation:
-
-            - At the end of the setup phase: ``post_setup``
-            - Every time step:
-              - ``time_step__prepare``
-              - ``time_step``
-              - ``time_step__cleanup``
-              - ``collect_metrics``
-            - At simulation end: ``simulation_end``
-
-        Parameters
-        ----------
-        event_name
-            The name of the event to listen for.
-        listener
-            The callable to be invoked any time an :class:`Event` with the given
-            name is emitted.
-        priority
-            One of {0, 1, 2, 3, 4, 5, 6, 7, 8, 9}.
-            An indication of the order in which event listeners should be
-            called. Listeners with smaller priority values will be called
-            earlier. Listeners with the same priority have no guaranteed
-            ordering.  This feature should be avoided if possible. Components
-            should strive to obey the Markov property as they transform the
-            state table (the state of the simulation at the beginning of the
-            next time step should only depend on the current state of the
-            system).
-        """
-        self._manager.register_listener(event_name, listener, priority)
