+++ src/vivarium/framework/randomness/manager.py
@@ -1,7 +1,7 @@
 """
-=========================
-Randomness System Manager
-=========================
+==================
+Randomness Manager
+==================
 
 """
