+++ src/vivarium/framework/results/manager.py
@@ -1,7 +1,7 @@
 """
-======================
-Results System Manager
-======================
+===============
+Results Manager
+===============
 
 """
