+++ src/vivarium/framework/values/manager.py
@@ -1,7 +1,7 @@
 """
-=====================
-Values System Manager
-=====================
+==============
+Values Manager
+==============
 
 """
